pyproject.toml
[project]
name = "cyrene"
version = "0.1.0"
description = "Cyrene AI Agent — self-evolving AI companion with SOUL.md memory"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.10,<4.0",
    "croniter>=6.0.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "requests>=2.34.0",
    "simplexng>=0.1.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cyrene"]

[tool.ruff]
line-length = 180
